jobs: temporary song downloading fix
don't pass cookies to yt-dlp, and remove the earlier player-client=tv workaround. youtube downloading is very in flux with available clients and what restrictions they have. newest version (2025.10.22) includes a temporary workaround to fix downloading, but no formats are available with cookies. hopefully this is resolved there soon.

the removal of cookies being sent to yt-dlp means certain age- and region-restricted videos will not be downloadable